requirements.txt
hydra-core==1.2.0
imageio==2.16.1
numpy==1.20.3
omegaconf==2.2.2
open3d==0.15.2
opencv_python==4.5.5.62
requests==2.26.0
scipy==1.8.1
setuptools==58.0.4
tensorboardX==2.5.1
tqdm
lpips
